requirements.txt
pandas
numpy
pygame
pytorch
pytest